chore: replace landing page with new design
Move new-landing-page to / and delete old landing page.
